feat: update github greetings readme file
Adds links to Google Fonts and a stylesheet in the README file. It also includes changes to the greetings section, adding a paragraph about the developer's background and interests, as well as a section for displaying badges with social media links. Additionally, it adds CSS styles for the greetings section, GitHub stats, and "Buy me a coffee" button.